WIP Linux port:  - All hardcoded paths converted to Unix format  - Fixes for a few Linux related crashes
Former-commit-id: 22215703c6336bef8e6fdc9c3a98a0551abf34c5

+++ Source/BansheeOISInput/BsInputHandlerOIS.h
@@ -51,7 +51,7 @@ namespace bs
 		};
 
 	public:
-		InputHandlerOIS(unsigned int hWnd);
+		InputHandlerOIS(UINT64 hWnd);
 		virtual ~InputHandlerOIS();
 
 	private:
